+++ src/test/java/com/clova/anifriends/domain/notification/vo/NotificationContentTest.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.notification.wrapper;
+package com.clova.anifriends.domain.notification.vo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
